chatbot-ui/requirements.txt
openai 
langchain 
tiktoken 
faiss-cpu 
python-dotenv 
pinecone-client
flask 
flask-cors
duckduckgo-search